+++ libraries/helpers.rb
@@ -9,14 +9,9 @@ module Gocd
     def get_agent_properties
       values = {}
       values[:go_server_url] = node['gocd']['agent']['go_server_url']
-      values[:go_server_port] = node['gocd']['agent']['go_server_port']
       if Chef::Config['solo'] || node['gocd']['agent']['go_server_url']
         Chef::Log.info("Attempting to use node['gocd']['agent']['go_server_url'] attribute for server url")
         values[:go_server_url] = node['gocd']['agent']['go_server_url']
-        values[:key] = node['gocd']['agent']['autoregister']['key']
-      elsif Chef::Config['solo'] || node['gocd']['agent']['go_server_host']
-        Chef::Log.info("Attempting to use node['gocd']['agent']['go_server_host'] attribute for server host")
-        values[:go_server_host] = node['gocd']['agent']['go_server_host']
         values[:key] = node['gocd']['agent']['autoregister']['key']
       else
         server_search_query = node['gocd']['agent']['server_search_query']
@@ -27,6 +22,7 @@ module Gocd
         else
           go_server = go_servers.first
           values[:go_server_host] = go_server['ipaddress']
+          values[:go_server_ssl_port] = go_server['gocd']['server']['https_port']
           if go_servers.count > 1
             Chef::Log.warn("Multiple Go servers found on Chef server. Using first returned server '#{values[:go_server_host]}' for server instance configuration.")
           end
@@ -34,7 +30,7 @@ module Gocd
           if values[:key] = go_server['gocd']['server']['autoregister_key']
             Chef::Log.warn("Agent auto-registration enabled. This agent will not require approval to become active.")
           end
-          values[:go_server_url] = "https://#{go_server}:#{values[:go_server_port]}/go"
+          values[:go_server_url] = "https://#{values[:go_server_host]}:#{values[:go_server_ssl_port]}/go"
         end
       end
       values[:hostname]     = node['gocd']['agent']['autoregister']['hostname']